ai-agent/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
arrow==1.4.0
asttokens==3.0.1
babel==2.17.0
bleach==6.3.0
click==8.3.1
dnspython==2.8.0
docx2txt==0.9
email-validator==2.3.0
faiss-cpu==1.13.0
fastapi==0.122.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastar==0.8.0
filelock==3.20.0
fqdn==1.5.1
fsspec==2025.10.0
hf-xet==1.2.0
httptools==0.7.1
huggingface-hub==0.36.0
ipykernel==7.1.0
isoduration==20.11.0
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jupyterlab==4.5.0
lark==1.3.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
nltk==3.9.2
numpy==2.3.5
pandas==2.3.3
pillow==12.0.0
pydantic==2.12.4
pydantic_core==2.41.5
PyMuPDF==1.26.6
pypandoc==1.16.2
pytesseract==0.3.13
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
rank-bm25==0.2.2
regex==2025.11.3
rfc3987-syntax==1.1.0
rich==14.2.0
rich-toolkit==0.16.0
rignore==0.7.6
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
Send2Trash==1.8.3
sentence-transformers==5.1.2
sentry-sdk==2.46.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
typing-inspection==0.4.2
tzdata==2025.2
uri-template==1.3.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
